backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-openai
langchain-google-genai
langchain-community
langchain-chroma
langchain-huggingface
sentence-transformers
chromadb
pypdf
tiktoken
python-dotenv
python-jose[cryptography]
bcrypt
sqlalchemy
rank_bm25
langchain-groq
